+++ src/components/form/checkbox/index.ts
@@ -1,0 +1,2 @@
+export { EuiCheckbox, EuiCheckboxProps } from './checkbox';
+export { EuiCheckboxGroup, EuiCheckboxGroupProps } from './checkbox_group';